- Added secondary particle effects, so when the main particle effect collides with an invisible plane in the scene, representing the SwiftUI GroupBox, a secondary particle effect activates. - Removed some dependencies and cleaned up LoadableLabel, TemperatureGauge. - TemperatureGauge transition is now properly clipped between the two overlays, so the transition doesn't extend beyond the frame of the gauge. - Added timeout to URLSession for `timeoutIntervalForResource`. - Removed unneeded WeatherConnection functions. - Fixed JSONDecoder dateDecodingStrategy so the incoming year from OpenWeather is correct. - Added more mock CurrentWeather functions for #Preview views. - Added a connection between LocationCard and particle view so the scroll offset of the scroll view can be used by the particle view to offset the plane that the particle effects collide with. - Combined async calls in LocationCard.downloadData by only calling await on all three download calls together (let (currentWeather, forecast, pollution) = try await (weatherDownload, forecastDownload, pollutionDownload) ) - Added landscape support to app, including particle effect views. - Decreased size of snowflake images. - Added new custom view to TransparentGroupBox so I can add a highlight to the view that aligns with ParticleEffectView.plane1.  Trying to give a slight dimensionality to the view so the particle effects appear more interactive with the overlay view. - Moved seperate views that were contained in a single file to their own files.

+++ Whether/Data Objects/CurrentWeather+Conditions.swift
@@ -9,7 +9,7 @@ import Foundation
 
 extension CurrentWeather {
     struct WeatherConditions: Identifiable, Codable, Equatable, Hashable, CustomStringConvertible {
-        let id: Double
+        let id: Int
         let mainLabel: String
         let weatherDetails: String
         let icon: String
@@ -37,7 +37,7 @@ extension CurrentWeather {
             URL(string: "https://openweathermap.org/img/wn/\(self.icon)@2x.png")!
         }
 
-        init(id: Double, mainLabel: String, description: String, icon: String) {
+        init(id: Int, mainLabel: String, description: String, icon: String) {
             self.id = id
             self.mainLabel = mainLabel
             self.weatherDetails = description
@@ -46,7 +46,7 @@ extension CurrentWeather {
 
         init(from decoder: any Decoder) throws {
             let container = try decoder.container(keyedBy: CodingKeys.self)
-            let id = try container.decode(Double.self, forKey: WeatherConditions.CodingKeys.id)
+            let id = try container.decode(Int.self, forKey: WeatherConditions.CodingKeys.id)
             self.id = id
             self.mainLabel = try container.decode(String.self, forKey: CodingKeys.mainLabel)
             self.weatherDetails = try container.decode(String.self, forKey: CodingKeys.weatherDetails)
